Added FATFS library. Verified microSD socket functionality and added aiSDInit() function for initializing an inserted card and the FATFS filesystem.

+++ AithonLibrary/ChibiOS/os/various/fatfs_bindings/fatfs_diskio.c
@@ -112,7 +112,7 @@ DRESULT disk_read (
     BYTE drv,        /* Physical drive nmuber (0..) */
     BYTE *buff,        /* Data buffer to store read data */
     DWORD sector,    /* Sector address (LBA) */
-    BYTE count        /* Number of sectors to read (1..255) */
+    UINT count        /* Number of sectors to read (1..255) */
 )
 {
   switch (drv) {
@@ -153,7 +153,7 @@ DRESULT disk_write (
     BYTE drv,            /* Physical drive nmuber (0..) */
     const BYTE *buff,    /* Data to be written */
     DWORD sector,        /* Sector address (LBA) */
-    BYTE count            /* Number of sectors to write (1..255) */
+    UINT count            /* Number of sectors to write (1..255) */
 )
 {
   switch (drv) {
